Damages two bug fixes and tests #464 #465

+++ app/routes/$lang+/disaster-record+/edit-sub.$disRecId+/damages+/delete.$id.tsx
@@ -16,7 +16,7 @@ import { getCountryAccountsIdFromSession } from "~/utils/session";
 export const action: ActionFunction = async (args) => {
 	const { request } = args;
 	const userSession = await requireUser(args);
-	
+
 	if (!userSession) {
 		throw new Response("Unauthorized", { status: 401 });
 	}
@@ -28,15 +28,17 @@ export const action: ActionFunction = async (args) => {
 
 	return createDeleteActionWithCountryAccounts({
 		redirectToSuccess: (_id: string, oldRecord: any) =>
-		 	route2(oldRecord.recordId) + "?sectorId=" + oldRecord.sectorId,
+			route2(oldRecord.recordId) + "?sectorId=" + oldRecord.sectorId,
 		delete: async (id: string) => {
 			return damagesDeleteById(id, countryAccountsId);
 		},
 		tableName: getTableName(damagesTable),
 		getById: damagesById,
 		postProcess: async (_id, data) => {
-			ContentRepeaterUploadFile.delete(data.attachments);
+			if (data.attachments) {
+				ContentRepeaterUploadFile.delete(data.attachments);
+			}
 		},
 		countryAccountsId,
 	})(args);
-}
+};
